requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Markdown==3.7
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
